Contained TaskPlugins in separate packages
