+++ flytekit/annotated/task.py
@@ -2,14 +2,10 @@ import collections
 import datetime as _datetime
 import functools
 import inspect
-import os
 import re
-import sys
 from abc import abstractmethod
 from enum import Enum
 from typing import Any, Callable, Dict, Generic, List, Optional, Tuple, Type, TypeVar, Union
-
-from google.protobuf.json_format import MessageToDict
 
 from flytekit.annotated.context_manager import BranchEvalMode, ExecutionState, FlyteContext, FlyteEntities
 from flytekit.annotated.interface import (
@@ -31,7 +27,6 @@ from flytekit.models import interface as _interface_models
 from flytekit.models import literals as _literal_models
 from flytekit.models import task as _task_model
 from flytekit.models.core import identifier as _identifier_model
-from flytekit.sdk.spark_types import SparkType
 
 
 def kwtypes(**kwargs) -> Dict[str, Type]:
@@ -441,91 +436,6 @@ class PythonFunctionTask(PythonTask, Generic[T]):
         )
 
 
-class Spark(object):
-    def __init__(self, spark_conf: Dict[str, str] = None, hadoop_conf: Dict[str, str] = None):
-        self._spark_conf = spark_conf
-        self._hadoop_conf = hadoop_conf
-
-    @property
-    def spark_conf(self) -> Dict[str, str]:
-        return self._spark_conf
-
-    @property
-    def hadoop_config(self) -> Dict[str, str]:
-        return self._hadoop_conf
-
-
-class GlobalSparkContext(object):
-    """
-    TODO revisit this class - it has been copied over from the spark task to avoid cyclic imports
-    """
-
-    _SPARK_CONTEXT = None
-    _SPARK_SESSION = None
-
-    @classmethod
-    def get_spark_context(cls):
-        return cls._SPARK_CONTEXT
-
-    @classmethod
-    def get_spark_session(cls):
-        return cls._SPARK_SESSION
-
-    def __enter__(self):
-        import pyspark as _pyspark
-
-        GlobalSparkContext._SPARK_CONTEXT = _pyspark.SparkContext()
-        GlobalSparkContext._SPARK_SESSION = _pyspark.sql.SparkSession.builder.appName(
-            "Flyte Spark SQL Context"
-        ).getOrCreate()
-        return self
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        GlobalSparkContext._SPARK_CONTEXT.stop()
-        GlobalSparkContext._SPARK_CONTEXT = None
-        return False
-
-
-class PysparkFunctionTask(PythonFunctionTask[Spark]):
-    def __init__(
-        self, task_config: Spark, task_function: Callable, metadata: _task_model.TaskMetadata, *args, **kwargs
-    ):
-        super(PysparkFunctionTask, self).__init__(
-            task_config=task_config,
-            task_type="spark",
-            task_function=task_function,
-            metadata=metadata,
-            ignore_input_vars=["spark_session", "spark_context"],
-            *args,
-            **kwargs,
-        )
-
-    def get_custom(self) -> Dict[str, Any]:
-        from flytekit.bin import entrypoint as _entrypoint
-
-        spark_exec_path = os.path.abspath(_entrypoint.__file__)
-        if spark_exec_path.endswith(".pyc"):
-            spark_exec_path = spark_exec_path[:-1]
-
-        job = _task_model.SparkJob(
-            spark_conf=self.task_config.spark_conf,
-            hadoop_conf=self.task_config.hadoop_config,
-            application_file="local://" + spark_exec_path,
-            executor_path=sys.executable,
-            main_class="",
-            spark_type=SparkType.PYTHON,
-        )
-        return MessageToDict(job.to_flyte_idl())
-
-    def execute(self, **kwargs) -> Any:
-        with GlobalSparkContext() as spark_ctx:
-            if "spark_session" in self.native_interface.inputs:
-                kwargs["spark_session"] = spark_ctx.get_spark_session()
-            if "spark_context" in self.native_interface.inputs:
-                kwargs["spark_context"] = spark_ctx.get_spark_context()
-            return self._task_function(**kwargs)
-
-
 class MapPythonTask(PythonTask):
     """
     TODO We might need a special entrypoint to start execution of this task type as there is possibly no instance of this
@@ -884,7 +794,6 @@ dynamic = functools.partial(task, task_config=_Dynamic())
 
 
 def _load_default_plugins():
-    TaskPlugins.register_pythontask_plugin(Spark, PysparkFunctionTask)
     TaskPlugins.register_pythontask_plugin(_Dynamic, DynamicWorkflowTask)
     TaskPlugins.register_pythontask_plugin(Reference, ReferenceTask)
