+++ tests/flytekit/unit/annotated/test_type_hints.py
@@ -10,7 +10,8 @@ from flytekit.annotated import context_manager, launch_plan, promise
 from flytekit.annotated.condition import conditional
 from flytekit.annotated.context_manager import ExecutionState
 from flytekit.annotated.promise import Promise
-from flytekit.annotated.task import ContainerTask, Reference, Spark, SQLTask, dynamic, kwtypes, maptask, metadata, task
+from flytekit.annotated.task import ContainerTask, Reference, SQLTask, dynamic, kwtypes, maptask, metadata, task
+from flytekit.taskplugins.spark.task import Spark
 from flytekit.annotated.testing import task_mock
 from flytekit.annotated.type_engine import RestrictedTypeError, TypeEngine
 from flytekit.annotated.workflow import workflow
